+++ packages/google-cloud-bigquery-analyticshub/google/cloud/bigquery_analyticshub_v1/services/analytics_hub_service/transports/grpc_asyncio.py
@@ -699,8 +699,8 @@ class AnalyticsHubServiceGrpcAsyncIOTransport(AnalyticsHubServiceTransport):
     ]:
         r"""Return a callable for the subscribe data exchange method over gRPC.
 
-        Creates a Subscription to a Data Exchange. This is a
-        long-running operation as it will create one or more
+        Creates a Subscription to a Data Clean Room. This is
+        a long-running operation as it will create one or more
         linked datasets.
 
         Returns: