+++ packages/google-cloud-compute/tests/unit/gapic/compute_v1/test_region_backend_services.py
@@ -5651,6 +5651,7 @@ def test_insert_rest_call_success(request_type):
             "failover_ratio": 0.1494,
         },
         "fingerprint": "fingerprint_value",
+        "ha_policy": {"fast_i_p_move": "fast_i_p_move_value"},
         "health_checks": ["health_checks_value1", "health_checks_value2"],
         "iap": {
             "enabled": True,
@@ -6338,6 +6339,7 @@ def test_patch_rest_call_success(request_type):
             "failover_ratio": 0.1494,
         },
         "fingerprint": "fingerprint_value",
+        "ha_policy": {"fast_i_p_move": "fast_i_p_move_value"},
         "health_checks": ["health_checks_value1", "health_checks_value2"],
         "iap": {
             "enabled": True,
@@ -7472,6 +7474,7 @@ def test_update_rest_call_success(request_type):
             "failover_ratio": 0.1494,
         },
         "fingerprint": "fingerprint_value",
+        "ha_policy": {"fast_i_p_move": "fast_i_p_move_value"},
         "health_checks": ["health_checks_value1", "health_checks_value2"],
         "iap": {
             "enabled": True,