+++ packages/google-cloud-dataplex/google/cloud/dataplex_v1/services/dataplex_service/transports/__init__.py
@@ -19,14 +19,18 @@ from typing import Dict, Type
 from .base import DataplexServiceTransport
 from .grpc import DataplexServiceGrpcTransport
 from .grpc_asyncio import DataplexServiceGrpcAsyncIOTransport
+from .rest import DataplexServiceRestInterceptor, DataplexServiceRestTransport
 
 # Compile a registry of transports.
 _transport_registry = OrderedDict()  # type: Dict[str, Type[DataplexServiceTransport]]
 _transport_registry["grpc"] = DataplexServiceGrpcTransport
 _transport_registry["grpc_asyncio"] = DataplexServiceGrpcAsyncIOTransport
+_transport_registry["rest"] = DataplexServiceRestTransport
 
 __all__ = (
     "DataplexServiceTransport",
     "DataplexServiceGrpcTransport",
     "DataplexServiceGrpcAsyncIOTransport",
+    "DataplexServiceRestTransport",
+    "DataplexServiceRestInterceptor",
 )