+++ packages/google-cloud-discoveryengine/google/cloud/discoveryengine_v1/services/conversational_search_service/transports/base.py
@@ -171,6 +171,11 @@ class ConversationalSearchServiceTransport(abc.ABC):
                 default_timeout=None,
                 client_info=client_info,
             ),
+            self.stream_answer_query: gapic_v1.method.wrap_method(
+                self.stream_answer_query,
+                default_timeout=None,
+                client_info=client_info,
+            ),
             self.get_answer: gapic_v1.method.wrap_method(
                 self.get_answer,
                 default_timeout=None,
@@ -289,6 +294,18 @@ class ConversationalSearchServiceTransport(abc.ABC):
 
     @property
     def answer_query(
+        self,
+    ) -> Callable[
+        [conversational_search_service.AnswerQueryRequest],
+        Union[
+            conversational_search_service.AnswerQueryResponse,
+            Awaitable[conversational_search_service.AnswerQueryResponse],
+        ],
+    ]:
+        raise NotImplementedError()
+
+    @property
+    def stream_answer_query(
         self,
     ) -> Callable[
         [conversational_search_service.AnswerQueryRequest],