+++ packages/google-cloud-discoveryengine/google/cloud/discoveryengine_v1/services/grounded_generation_service/async_client.py
@@ -382,7 +382,7 @@ class GroundedGenerationServiceAsyncClient:
 
         Returns:
             AsyncIterable[google.cloud.discoveryengine_v1.types.GenerateGroundedContentResponse]:
-
+                Response for the GenerateGroundedContent method.
         """
 
         # Wrap the RPC method; this adds retry and timeout information,
@@ -461,7 +461,7 @@ class GroundedGenerationServiceAsyncClient:
 
         Returns:
             google.cloud.discoveryengine_v1.types.GenerateGroundedContentResponse:
-
+                Response for the GenerateGroundedContent method.
         """
         # Create or coerce a protobuf request object.
         # - Use the request object if provided (there's no risk of modifying the input as