+++ packages/google-cloud-discoveryengine/google/cloud/discoveryengine_v1/services/grounded_generation_service/transports/rest.py
@@ -576,7 +576,7 @@ class GroundedGenerationServiceRestTransport(
 
             Returns:
                 ~.grounded_generation_service.GenerateGroundedContentResponse:
-
+                    Response for the ``GenerateGroundedContent`` method.
             """
 
             http_options = (