+++ packages/google-cloud-discoveryengine/google/cloud/discoveryengine_v1/types/user_event_service.py
@@ -42,8 +42,9 @@ class WriteUserEventRequest(proto.Message):
             [DataStore][google.cloud.discoveryengine.v1.DataStore]
             level, the format is:
             ``projects/{project}/locations/{location}/collections/{collection}/dataStores/{data_store}``.
-            If the write user event action is applied in [Location][]
-            level, for example, the event with
+            If the write user event action is applied in
+            [Location][google.cloud.location.Location] level, for
+            example, the event with
             [Document][google.cloud.discoveryengine.v1.Document] across
             multiple
             [DataStore][google.cloud.discoveryengine.v1.DataStore], the
@@ -81,8 +82,18 @@ class CollectUserEventRequest(proto.Message):
 
     Attributes:
         parent (str):
-            Required. The parent DataStore resource name, such as
+            Required. The parent resource name. If the collect user
+            event action is applied in
+            [DataStore][google.cloud.discoveryengine.v1.DataStore]
+            level, the format is:
             ``projects/{project}/locations/{location}/collections/{collection}/dataStores/{data_store}``.
+            If the collect user event action is applied in
+            [Location][google.cloud.location.Location] level, for
+            example, the event with
+            [Document][google.cloud.discoveryengine.v1.Document] across
+            multiple
+            [DataStore][google.cloud.discoveryengine.v1.DataStore], the
+            format is: ``projects/{project}/locations/{location}``.
         user_event (str):
             Required. URL encoded UserEvent proto with a
             length limit of 2,000,000 characters.